fix old tests
